Game Board and Chips
Inside App.css there is code for red and yellow chips.

The board is just a simple square. In the future we could consider making the board have a see through hole so that the background of the website is visible but if we use a simple solid color background it wont be needed.