(Read description) Fixed Victory Check
The issue was that game.symbol was being passed in as the opposite value that we wanted it to be.

For example:
blue chip = false
yellow chip = true

in victory check, blue chip would count for trues instead of for false and vice versa.

to fix the issue, we simply added a NOT operator in front of game.symbol in line 73 such that it would equal the opposite of what it was meant too.

Alternatively, you can make it so that blue chip is represented as true and yellow as false

I also added some other debugging logs inside victory check in order to find the source of the problem